Restrict customer actions to superadmins with FirebaseAuthGuard
Added FirebaseAuthGuard to secure customer endpoints and restricted access to superadmins only. Updated the CustomersController methods to enforce this restriction and handle unauthorized access with ForbiddenException. Modified CustomersModule to include necessary dependencies like AuthService and UsersService.

+++ src/customers/customers.controller.ts
@@ -7,28 +7,49 @@ import {
   Param,
   Delete,
   Query,
+  UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
 import { CustomersService } from './customers.service';
 import { CreateCustomerDto } from './dto/create-customer.dto';
 import { UpdateCustomerDto } from './dto/update-customer.dto';
 import { ListCustomersInputDto } from './dto/list-customers-input.dto';
+import { FirebaseAuthGuard } from '../auth/guards/firebase-auth/firebase-auth.guard';
+import { User } from '../common/decorators/user.decorator';
+import { OutputUserDto } from '../users/dto/output-user.dto';
 
 @Controller('customers')
+@UseGuards(FirebaseAuthGuard)
 export class CustomersController {
   constructor(private readonly customersService: CustomersService) {}
 
   @Post()
-  create(@Body() createCustomerDto: CreateCustomerDto) {
+  create(
+    @User() user: OutputUserDto,
+    @Body() createCustomerDto: CreateCustomerDto,
+  ) {
+    if (!user.isSuperadmin) {
+      throw new ForbiddenException('You have no access to customers.');
+    }
     return this.customersService.create(createCustomerDto);
   }
 
   @Get()
-  findAll(@Query() listCustomersInputDto: ListCustomersInputDto) {
+  findAll(
+    @Query() listCustomersInputDto: ListCustomersInputDto,
+    @User() user: OutputUserDto,
+  ) {
+    if (!user.isSuperadmin) {
+      throw new ForbiddenException('You have no access to customers.');
+    }
     return this.customersService.findAll(listCustomersInputDto);
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  findOne(@Param('id') id: string, @User() user: OutputUserDto) {
+    if (!user.isSuperadmin) {
+      throw new ForbiddenException('You have no access to customers.');
+    }
     return this.customersService.findOne(+id);
   }
 
@@ -36,12 +57,19 @@ export class CustomersController {
   update(
     @Param('id') id: string,
     @Body() updateCustomerDto: UpdateCustomerDto,
+    @User() user: OutputUserDto,
   ) {
+    if (!user.isSuperadmin) {
+      throw new ForbiddenException('You have no access to customers.');
+    }
     return this.customersService.update(+id, updateCustomerDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Param('id') id: string, @User() user: OutputUserDto) {
+    if (!user.isSuperadmin) {
+      throw new ForbiddenException('You have no access to customers.');
+    }
     return this.customersService.remove(+id);
   }
 }
